#288 refactor(component): Update import paths for BottomNavItem and ConfigChildComponent
This commit updates the import paths for `BottomNavItem` and `ConfigChildComponent` in the `Component.kt` file.

The previous imports from the `com.rahim.yadino.navigation` package have been replaced with new paths pointing to `com.rahim.component`. This change aligns the imports with a recent refactoring or relocation of these components.

**Key Changes:**

*   **`Component.kt`**:
    *   Changed import from `com.rahim.yadino.navigation.BottomNavItem` to `com.rahim.component.BottomNavItem`.
    *   Changed import from `com.rahim.yadino.navigation.config.ConfigChildComponent` to `com.rahim.component.config.ConfigChildComponent`.